==> notebooks/01_districts_centroids.ipynb <==
import geopandas as gpd
import pandas as pd

import geopandas as gpd

shp_path = "../data/raw/admin_boundaries/gadm41_TUR_shp/gadm41_TUR_2.shp"

gdf = gpd.read_file(shp_path, engine="fiona")  # fiona engine ile açıyoruz
gdf.head()

gdf = gdf.to_crs(epsg=4326)
gdf.crs


gdf.columns


gdf_simple = gdf[["NAME_1", "NAME_2", "geometry"]].copy()
gdf_simple = gdf_simple.rename(columns={
    "NAME_1": "province_name",
    "NAME_2": "district_name"
})
gdf_simple.head()


gdf_simple["centroid"] = gdf_simple.geometry.centroid
gdf_simple["lat"] = gdf_simple.centroid.y
gdf_simple["lon"] = gdf_simple.centroid.x
gdf_simple.head()

# Başındaki/sonundaki boşlukları temizle
gdf_simple["province_name"] = gdf_simple["province_name"].str.strip()
gdf_simple["district_name"] = gdf_simple["district_name"].str.strip()

# İstersen hepsini Title Case yap
gdf_simple["province_name"] = gdf_simple["province_name"].str.title()
gdf_simple["district_name"] = gdf_simple["district_name"].str.title()


df_out = gdf_simple[["province_name", "district_name", "lat", "lon"]].copy()

# Kaç ilçe var bakalım
len(df_out)


import os

# ../data/interim klasörünü oluştur (varsa hata vermez)
os.makedirs("../data/interim", exist_ok=True)


output_path = "../data/interim/districts_base.csv"
df_out.to_csv(output_path, index=False)

len(df_out), output_path
