==> notebooks/02_climate_aggregation.ipynb <==
import pandas as pd
import os


csv_path = "../data/raw/climate_nasa_power/districts_monthly_2014_2023.csv"

df = pd.read_csv(csv_path)
df.head(), df.shape


df.dtypes


# Gruplama anahtarları: il, ilçe, koordinat, yıl
group_cols = ["province_name", "district_name", "lat", "lon", "year"]

yearly = (
    df
    .groupby(group_cols, as_index=False)
    .agg(
        mean_temp_year=("t2m", "mean"),          # o yılın ortalama sıcaklığı
        total_precip_year=("prectot", "sum"),    # o yılın toplam yağışı
        min_temp_year=("t2m", "min"),
        max_temp_year=("t2m", "max"),
    )
)

yearly.head(), yearly.shape


os.makedirs("../data/interim", exist_ok=True)


yearly_path = "../data/interim/districts_climate_yearly_2014_2023.csv"
yearly.to_csv(yearly_path, index=False)
yearly_path


ten_year = (
    yearly
    .groupby(["province_name", "district_name", "lat", "lon"], as_index=False)
    .agg(
        avg_temp_10yr=("mean_temp_year", "mean"),
        avg_precip_10yr=("total_precip_year", "mean"),
        temp_std_10yr=("mean_temp_year", "std"),
        precip_std_10yr=("total_precip_year", "std"),
        temp_min_10yr=("mean_temp_year", "min"),
        temp_max_10yr=("mean_temp_year", "max"),
    )
)

ten_year.head(), ten_year.shape


ten_year_path = "../data/interim/districts_climate_10yr.csv"
ten_year.to_csv(ten_year_path, index=False)
ten_year_path
